+++ packages/bezier-react/src/components/Stack/Stack/Stack.types.ts
@@ -54,4 +54,5 @@ interface StackOptions {
 export default interface StackProps extends
   BezierComponentProps,
   ChildrenProps,
+  React.HTMLAttributes<HTMLElement>,
   StackOptions {}