+++ packages/bezier-react/src/layout/components/HeaderArea/HeaderArea.types.ts
@@ -3,6 +3,7 @@ import { BezierComponentProps } from 'Types/ComponentProps'
 
 interface HeaderAreaOptions {
   hasHeader: boolean
+  showSideView: boolean
   ContentHeaderComponent: React.ComponentType
   CoverableHeaderComponent?: React.ComponentType
 }